Fixed iterations, added breaks
This will likely only work for easy to medium sodokus.  Hard sodokus require assumptions sometimes which this program does not cover.
I added a condition to the loop to break it out out after so many seconds.  Each iteration takes a tiny amount of time.  So if you put a really hard puzzle in, it will at least break out of the loop and give you back whatever it did figure out (I hope).  This has not been tested.